+++ examples/packages/project-utils/packages/index.js
@@ -1,1 +1,3 @@
-module.exports = require("get-yarn-workspaces")();
+const path = require("path");
+
+module.exports = require("get-yarn-workspaces")().map(pkg => pkg.replace(/\//g, path.sep));