+++ packages/api-cms/src/plugins/models.js
@@ -1,19 +1,26 @@
 // @flow
 import { flow } from "lodash";
 import { withUser } from "@webiny/api-security";
-import { withStorage, withCrudLogs, withSoftDelete } from "@webiny/commodo";
-import { MongoDbDriver, withId } from "@commodo/fields-storage-mongodb";
+import { withStorage, withCrudLogs, withSoftDelete, withFields } from "@webiny/commodo";
 import cmsContent from "./models/cmsContent.model";
 
-export default ({ database }) => ({
+export default () => ({
     name: "graphql-context-models",
     type: "graphql-context",
     apply(context) {
-        const driver = new MongoDbDriver({ database });
+        const driver = context.commodo && context.commodo.driver;
+
+        if (!driver) {
+            throw Error(
+                `Commodo driver is not configured! Make sure you add a Commodo driver plugin to your service.`
+            );
+        }
 
         const createBase = () =>
             flow(
-                withId(),
+                withFields({
+                    id: context.commodo.fields.id()
+                }),
                 withStorage({ driver }),
                 withUser(context),
                 withSoftDelete(),