+++ packages/api-files/src/plugins/graphql.js
@@ -1,5 +1,5 @@
 // @flow
-import { gql } from "apollo-server-lambda";
+import gql from "graphql-tag";
 import { emptyResolver } from "@webiny/commodo-graphql";
 import { hasScope } from "@webiny/api-security";
 import { resolveCreate, resolveGet, resolveUpdate } from "@webiny/commodo-graphql";
@@ -117,6 +117,7 @@ export default ([
                         perPage: Int
                         types: [String]
                         tags: [String]
+                        ids: [ID]
                         sort: JSON
                         search: String
                     ): FileListResponse
@@ -189,7 +190,6 @@ export default ([
             shield: {
                 FilesQuery: {
                     getFile: hasScope("files:file:crud"),
-                    listFiles: hasScope("files:file:crud")
                 },
                 FilesMutation: {
                     uploadFile: hasScope("files:file:crud"),