+++ packages/api-files/src/plugins/resolvers/listTags.js
@@ -1,16 +1,10 @@
 // @flow
 export default async (root: any, args: Object, context: Object) => {
-    const { File } = context.models;
+    const plugin = context.plugins.byName("files-resolver-list-tags");
 
-    const results = await File.aggregate([
-        { $sort: { _id: 1 } },
-        { $match: { tags: { $exists: true, $ne: [] } } },
-        { $project: { tags: 1 } },
-        { $unwind: "$tags" },
-        { $group: { _id: "$tags" } },
-        { $sort: { _id: 1 } },
-        { $limit: 100 }
-    ]);
+    if (!plugin) {
+        throw Error(`Resolver plugin "files-resolver-list-tags" is not configured!`);
+    }
 
-    return results.map(item => item._id);
+    return await plugin.resolve({ context });
 };