+++ packages/api-page-builder/src/plugins/models.js
@@ -1,7 +1,6 @@
 // @flow
 import { flow } from "lodash";
-import { withStorage, withCrudLogs, withSoftDelete } from "@webiny/commodo";
-import { MongoDbDriver, withId } from "@commodo/fields-storage-mongodb";
+import { withStorage, withCrudLogs, withSoftDelete, withFields } from "@webiny/commodo";
 import { withUser } from "@webiny/api-security";
 
 import pbCategory from "./models/pbCategory.model";
@@ -10,23 +9,31 @@ import pbMenu from "./models/pbMenu.model";
 import pbPage from "./models/pbPage.model";
 import pbSettings from "./models/pbSettings.model";
 
-export default ({ database }) => ({
+export default () => ({
     name: "graphql-context-models",
     type: "graphql-context",
     apply(context) {
-        const driver = new MongoDbDriver({ database });
+        const driver = context.commodo && context.commodo.driver;
+
+        if (!driver) {
+            throw Error(
+                `Commodo driver is not configured! Make sure you add a Commodo driver plugin to your service.`
+            );
+        }
 
         const createBase = ({ maxPerPage = 100 } = {}) =>
             flow(
-                withId(),
+                withFields({
+                    id: context.commodo.fields.id()
+                }),
                 withStorage({ driver, maxPerPage }),
                 withUser(context),
                 withSoftDelete(),
                 withCrudLogs()
             )();
 
-        const PbCategory = pbCategory({ createBase });
-        const PbMenu = pbMenu({ createBase });
+        const PbCategory = pbCategory({ createBase, context });
+        const PbMenu = pbMenu({ createBase, context });
         const PbPageElement = pbPageElement({ createBase, context });
         const PbSettings = pbSettings({ createBase, context });
         const PbPage = pbPage({ createBase, context, PbCategory, PbSettings });