+++ components/serverless-files/CHANGELOG.md
@@ -3,6 +3,30 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [2.3.0-next.1](https://github.com/Pavel910/webiny-js/compare/@webiny/serverless-files@2.3.0-next.0...@webiny/serverless-files@2.3.0-next.1) (2019-11-20)
+
+**Note:** Version bump only for package @webiny/serverless-files
+
+
+
+
+
+# [2.3.0-next.0](https://github.com/Pavel910/webiny-js/compare/@webiny/serverless-files@2.2.0...@webiny/serverless-files@2.3.0-next.0) (2019-11-18)
+
+
+### Bug Fixes
+
+* make plugins an empty array by default ([b2eaa1f](https://github.com/Pavel910/webiny-js/commit/b2eaa1f))
+
+
+### Features
+
+* add support for DB drivers and improve components. ([3ce1908](https://github.com/Pavel910/webiny-js/commit/3ce1908))
+
+
+
+
+
 # [2.2.0](https://github.com/Webiny/webiny-js/compare/@webiny/serverless-files@2.1.0...@webiny/serverless-files@2.2.0) (2019-11-08)